Modularize pipeline editor and python code (#411)
Refactor Elyra  extension code to better modularize widgets
into its own TSX file and changes some of the class names to be
more clear / consistent.

Fixes #395
